issue #9291:  Fix typo in xtreewidget that was preventing higher precision numbers from being initialized.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@6618 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ widgets/xtreewidget.cpp
@@ -945,7 +945,7 @@ void XTreeWidget::populateCalculatedColumns()
       {
         int set = topLevelItem(row)->data(col, RunningSetRole).toInt();
         if (! subtotals.contains(set))
-          subtotals[set] = topLevelItem(row)->data(col, RunningInitRole).toInt();
+          subtotals[set] = topLevelItem(row)->data(col, RunningInitRole).toDouble();
         subtotals[set] += topLevelItem(row)->data(col, RawRole).toDouble();
         topLevelItem(row)->setData(col, Qt::DisplayRole,
                                    QLocale().toString(subtotals[set], 'f',
